Add helper function for window motions
This function allows me to seamlessly move around vim split windows and
tmux panes using the same binds.  If my tmux window is split into two
panes with vim in the top and a shell in the bottom, I would be able to
break out of vim with ^Wj and be in the shell window.

Additional configuration and a helper script is required for tmux to
allow this communication to happen.